Made the player controllers use inheritance so SpeedBuff.cs would still work
