+++ Assets/Scripts/PlayerMovement/MovementData_Gatherer.asset
@@ -12,9 +12,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d56ae2fa66cae494bbf52c33c6c92dc0, type: 3}
   m_Name: MovementData_Gatherer
   m_EditorClassIdentifier: 
-  maxSpeed: 15
-  acceleration: 5
-  deceleration: 2.58
-  accelerationForce: 16.666666
-  decelerationForce: 8.599999
+  maxSpeed: 7.5
+  acceleration: 1.5
+  deceleration: 2.5
+  accelerationForce: 10
+  decelerationForce: 16.666666
   conserveMomentum: 1
