+++ Assets/Scripts/PlayerMovement/MovementData_Warden.asset
@@ -12,9 +12,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d56ae2fa66cae494bbf52c33c6c92dc0, type: 3}
   m_Name: MovementData_Warden
   m_EditorClassIdentifier: 
-  maxSpeed: 20
-  acceleration: 1.51
-  deceleration: 2.58
-  accelerationForce: 3.775
-  decelerationForce: 6.45
+  maxSpeed: 12.5
+  acceleration: 1.5
+  deceleration: 2.5
+  accelerationForce: 6
+  decelerationForce: 10
   conserveMomentum: 0
